+++ libs/ethereum/index.ts
@@ -1,9 +1,12 @@
 import { Web3Websocket, isMetaMaskAvailable } from "./web3Websocket";
 import { getEthAccountAddress, getEthAccountFromMetaMask } from "./accounts";
+import { estimateGas, estimateGasPrice } from "./gas";
 
 export {
   Web3Websocket,
   getEthAccountAddress,
   isMetaMaskAvailable,
   getEthAccountFromMetaMask,
+  estimateGas,
+  estimateGasPrice,
 };