+++ libs/tokens/token.ts
@@ -1,21 +1,14 @@
-import path from "path";
 import type Web3 from "web3";
 import type { Contract } from "web3-eth-contract";
-import { parentLogger } from "../utils";
+import { logger, NightfallSdkError } from "../utils";
 import { APPROVE_AMOUNT } from "./constants";
 import type { TokenOptions } from "./types";
 import erc20Abi from "./abis/ERC20.json";
-import erc165Abi from "../../libs/tokens/abis/ERC165.json";
+import erc165Abi from "./abis/ERC165.json";
 import erc721Abi from "./abis/ERC721.json";
 import erc1155Abi from "./abis/ERC1155.json";
 import type { AbiItem } from "web3-utils";
-import { NightfallSdkError } from "../utils/error";
 import { ERC20, ERC721, ERC1155, ERC721_INTERFACE_ID } from "./constants";
-
-const logger = parentLogger.child({
-  name: path.relative(process.cwd(), __filename),
-});
-
 /**
  * Detects ERC standard for a given contract address using ERC165
  *