+++ libs/transactions/helpers/submit.ts
@@ -1,24 +1,11 @@
 import type Web3 from "web3";
-import path from "path";
-import { parentLogger } from "../../utils";
-import type { TransactionReceipt } from "web3-core";
-
-const logger = parentLogger.child({
-  name: path.relative(process.cwd(), __filename),
-});
-
-const GAS = process.env.GAS || 4000000;
-const GAS_PRICE = process.env.GAS_PRICE || 10000000000;
-// const GAS_ESTIMATE_ENDPOINT =
-//   process.env.GAS_ESTIMATE_ENDPOINT ||
-//   "https://vqxy02tr5e.execute-api.us-east-2.amazonaws.com/production/estimateGas";
-
-const GAS_MULTIPLIER = Number(process.env.GAS_MULTIPLIER) || 2;
-const GAS_PRICE_MULTIPLIER = Number(process.env.GAS_PRICE_MULTIPLIER) || 2;
+import type { TransactionConfig, TransactionReceipt } from "web3-core";
+import { logger, NightfallSdkError } from "../../utils";
+import { estimateGas, estimateGasPrice } from "../../ethereum";
 
 /**
  * Create, sign and broadcast an Ethereum transaction (tx) to the network
- * 
+ *
  * @async
  * @function submitTransaction
  * @param {string} senderEthAddress Eth address sending the contents of the tx
@@ -26,7 +13,7 @@ const GAS_PRICE_MULTIPLIER = Number(process.env.GAS_PRICE_MULTIPLIER) || 2;
  * @param {string} recipientEthAddress Eth address receiving the contents of the tx
  * @param {string} unsignedTx The contents of the tx (sent in data)
  * @param {Web3} web3 web3js instance
- * @param {string} fee The amount in Wei to pay a proposer for the tx
+ * @param {string} value Proposer payment for the tx in L1
  * @returns {Promise<TransactionReceipt>}
  */
 export async function submitTransaction(
@@ -35,36 +22,43 @@ export async function submitTransaction(
   recipientEthAddress: string,
   unsignedTx: string,
   web3: Web3,
-  fee = "0",
+  value = "0",
 ): Promise<TransactionReceipt> {
   logger.debug(
-    { senderEthAddress, recipientEthAddress, unsignedTx, fee },
+    { senderEthAddress, recipientEthAddress, unsignedTx, value },
     "submitTransaction",
   );
 
-  // Estimate gas
-  const gas = Math.ceil(Number(GAS) * GAS_MULTIPLIER); // ISSUE #28
-  const gasPrice = Math.ceil(Number(GAS_PRICE) * GAS_PRICE_MULTIPLIER); // ISSUE #28
-  logger.debug(
-    `Transaction gasPrice was set at ${Math.ceil(
-      gasPrice / 10 ** 9,
-    )} GWei, gas limit was set at ${gas}`,
-  );
+  const isListening = await web3.eth.net.isListening();
+  if (!isListening)
+    throw new NightfallSdkError(
+      "Web3 websocket not listening, try again later",
+    );
 
-  const tx = {
+  // Estimate gasPrice
+  const gasPrice = await estimateGasPrice(web3);
+
+  // Ethereum tx
+  const tx: TransactionConfig = {
     from: senderEthAddress,
     to: recipientEthAddress,
     data: unsignedTx,
-    value: fee,
-    gas,
+    value,
     gasPrice,
   };
 
+  // Estimate tx gas
+  const gas = await estimateGas(tx, web3);
+  logger.debug(`Transaction gas set at ${gas}`);
+  tx.gas = gas;
+
+  // If no private key is given, SDK tries to submit tx via MetaMask
   if (!senderEthPrivateKey) {
     logger.debug({ tx }, "Send tx via MetaMask...");
     return web3.eth.sendTransaction(tx);
   }
 
+  // Otherwise, sign tx then submit it
   logger.debug({ tx }, "Sign tx...");
   const signedTx = await web3.eth.accounts.signTransaction(
     tx,