+++ libs/utils/logger.ts
@@ -1,10 +1,13 @@
 import pino from "pino";
 
-const LOGGER_DEFAULT_LEVEL = process.env.LOG_LEVEL || "info";
+const LOGGER_DEFAULT_LEVEL = globalThis.process?.env.LOG_LEVEL ?? "info";
 const LOGGER_TIME_STRING = "yyyy-mm-dd HH:MM:ss";
 
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+// https://getpino.io/#/docs/transports?id=typescript-compatibility
 const logger = pino({
-  level: LOGGER_DEFAULT_LEVEL, // ISSUE #33
+  level: LOGGER_DEFAULT_LEVEL,
   transport: {
     target: "pino-pretty",
     options: {